# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for ClipIO and platform hub
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Caption/subtitle style
 */
export type ProcessVideoRequestCaptionStyle =
  (typeof ProcessVideoRequestCaptionStyle)[keyof typeof ProcessVideoRequestCaptionStyle];

export const ProcessVideoRequestCaptionStyle = {
  default: "default",
  hormozi: "hormozi",
  mrbeast: "mrbeast",
  minimal: "minimal",
  tiktok: "tiktok",
  neon: "neon",
  podcast: "podcast",
} as const;

export interface ProcessVideoRequest {
  /** YouTube video URL to process */
  youtubeUrl: string;
  /** Caption/subtitle style */
  captionStyle?: ProcessVideoRequestCaptionStyle;
  /**
   * Maximum number of clips to generate
   * @minimum 1
   * @maximum 10
   */
  maxClips?: number;
}

export interface ProcessVideoResponse {
  jobId: string;
  message: string;
  estimatedMinutes?: number;
}

export type JobStatusStatus =
  (typeof JobStatusStatus)[keyof typeof JobStatusStatus];

export const JobStatusStatus = {
  pending: "pending",
  downloading: "downloading",
  transcribing: "transcribing",
  analyzing: "analyzing",
  clipping: "clipping",
  completed: "completed",
  failed: "failed",
} as const;

export interface JobStatus {
  jobId: string;
  status: JobStatusStatus;
  /**
   * @minimum 0
   * @maximum 100
   */
  progress: number;
  message: string;
  videoTitle?: string;
  error?: string;
}

export interface ClipInfo {
  clipId: string;
  title: string;
  startTime: string;
  endTime: string;
  duration: number;
  viralityScore: number;
  hookType?: string;
  reasoning?: string;
  downloadUrl: string;
  thumbnailUrl?: string;
}

export interface ClipsResponse {
  jobId: string;
  videoTitle?: string;
  clips: ClipInfo[];
}

export interface ErrorResponse {
  error: string;
  details?: string;
}