# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for ClipIO and platform hub
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: clipio
    description: ClipIO video processing operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /clipio/process:
    post:
      operationId: processVideo
      tags: [clipio]
      summary: Start video processing
      description: Download YouTube video, transcribe, analyze for viral moments, and generate clips
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessVideoRequest"
      responses:
        "200":
          description: Processing started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessVideoResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /clipio/status/{jobId}:
    get:
      operationId: getJobStatus
      tags: [clipio]
      summary: Get job processing status
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStatus"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /clipio/clips/{jobId}:
    get:
      operationId: getClips
      tags: [clipio]
      summary: Get generated clips for a completed job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of generated clips
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClipsResponse"
        "404":
          description: Job not found or not completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ProcessVideoRequest:
      type: object
      properties:
        youtubeUrl:
          type: string
          description: YouTube video URL to process
        captionStyle:
          type: string
          enum: [default, hormozi, mrbeast, minimal, tiktok, neon, podcast]
          default: tiktok
          description: Caption/subtitle style
        maxClips:
          type: integer
          minimum: 1
          maximum: 10
          default: 5
          description: Maximum number of clips to generate
      required:
        - youtubeUrl

    ProcessVideoResponse:
      type: object
      properties:
        jobId:
          type: string
        message:
          type: string
        estimatedMinutes:
          type: number
      required:
        - jobId
        - message

    JobStatus:
      type: object
      properties:
        jobId:
          type: string
        status:
          type: string
          enum: [pending, downloading, transcribing, analyzing, clipping, completed, failed]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        message:
          type: string
        videoTitle:
          type: string
        error:
          type: string
      required:
        - jobId
        - status
        - progress
        - message

    ClipInfo:
      type: object
      properties:
        clipId:
          type: string
        title:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        duration:
          type: number
        viralityScore:
          type: integer
        hookType:
          type: string
        reasoning:
          type: string
        downloadUrl:
          type: string
        thumbnailUrl:
          type: string
      required:
        - clipId
        - title
        - startTime
        - endTime
        - duration
        - viralityScore
        - downloadUrl

    ClipsResponse:
      type: object
      properties:
        jobId:
          type: string
        videoTitle:
          type: string
        clips:
          type: array
          items:
            $ref: "#/components/schemas/ClipInfo"
      required:
        - jobId
        - clips

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      required:
        - error