# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for ClipIO and platform hub
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Download YouTube video, transcribe, analyze for viral moments, and generate clips
 * @summary Start video processing
 */
export const processVideoBodyCaptionStyleDefault = `tiktok`;
export const processVideoBodyMaxClipsDefault = 5;
export const processVideoBodyMaxClipsMax = 10;

export const ProcessVideoBody = zod.object({
  youtubeUrl: zod.string().describe("YouTube video URL to process"),
  captionStyle: zod
    .enum([
      "default",
      "hormozi",
      "mrbeast",
      "minimal",
      "tiktok",
      "neon",
      "podcast",
    ])
    .default(processVideoBodyCaptionStyleDefault)
    .describe("Caption\/subtitle style"),
  maxClips: zod
    .number()
    .min(1)
    .max(processVideoBodyMaxClipsMax)
    .default(processVideoBodyMaxClipsDefault)
    .describe("Maximum number of clips to generate"),
});

export const ProcessVideoResponse = zod.object({
  jobId: zod.string(),
  message: zod.string(),
  estimatedMinutes: zod.number().optional(),
});

/**
 * @summary Get job processing status
 */
export const GetJobStatusParams = zod.object({
  jobId: zod.coerce.string(),
});

export const getJobStatusResponseProgressMin = 0;
export const getJobStatusResponseProgressMax = 100;

export const GetJobStatusResponse = zod.object({
  jobId: zod.string(),
  status: zod.enum([
    "pending",
    "downloading",
    "transcribing",
    "analyzing",
    "clipping",
    "completed",
    "failed",
  ]),
  progress: zod
    .number()
    .min(getJobStatusResponseProgressMin)
    .max(getJobStatusResponseProgressMax),
  message: zod.string(),
  videoTitle: zod.string().optional(),
  error: zod.string().optional(),
});

/**
 * @summary Get generated clips for a completed job
 */
export const GetClipsParams = zod.object({
  jobId: zod.coerce.string(),
});

export const GetClipsResponse = zod.object({
  jobId: zod.string(),
  videoTitle: zod.string().optional(),
  clips: zod.array(
    zod.object({
      clipId: zod.string(),
      title: zod.string(),
      startTime: zod.string(),
      endTime: zod.string(),
      duration: zod.number(),
      viralityScore: zod.number(),
      hookType: zod.string().optional(),
      reasoning: zod.string().optional(),
      downloadUrl: zod.string(),
      thumbnailUrl: zod.string().optional(),
    }),
  ),
});