# lib/api-zod/src/generated/types/clipInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for ClipIO and platform hub
 * OpenAPI spec version: 0.1.0
 */

export interface ClipInfo {
  clipId: string;
  title: string;
  startTime: string;
  endTime: string;
  duration: number;
  viralityScore: number;
  hookType?: string;
  reasoning?: string;
  downloadUrl: string;
  thumbnailUrl?: string;
}