# lib/api-zod/src/generated/types/clipsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for ClipIO and platform hub
 * OpenAPI spec version: 0.1.0
 */
import type { ClipInfo } from "./clipInfo";

export interface ClipsResponse {
  jobId: string;
  videoTitle?: string;
  clips: ClipInfo[];
}