# lib/api-zod/src/generated/types/jobStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for ClipIO and platform hub
 * OpenAPI spec version: 0.1.0
 */
import type { JobStatusStatus } from "./jobStatusStatus";

export interface JobStatus {
  jobId: string;
  status: JobStatusStatus;
  /**
   * @minimum 0
   * @maximum 100
   */
  progress: number;
  message: string;
  videoTitle?: string;
  error?: string;
}