# lib/api-zod/src/generated/types/processVideoRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for ClipIO and platform hub
 * OpenAPI spec version: 0.1.0
 */
import type { ProcessVideoRequestCaptionStyle } from "./processVideoRequestCaptionStyle";

export interface ProcessVideoRequest {
  /** YouTube video URL to process */
  youtubeUrl: string;
  /** Caption/subtitle style */
  captionStyle?: ProcessVideoRequestCaptionStyle;
  /**
   * Maximum number of clips to generate
   * @minimum 1
   * @maximum 10
   */
  maxClips?: number;
}